Add files via upload
Polynomial regression is a type of regression analysis that is used to model relationships between a dependent variable and one or more independent variables using a polynomial equation. It is often used when data exhibits a non-linear relationship, as linear regression may not produce accurate results in such cases. In polynomial regression, the original features are transformed into polynomial features of a specified degree and modeled using a linear model.

The aim of this project is to build a polynomial regression model to predict the points scored by a sports team based on various attributes. The project will involve exploratory data analysis, data preprocessing, and model building, as well as a thorough explanation of various regression metrics.